# encoding: do not cache large encodeState