--- a/src/encoding/json/stream.go
+++ b/src/encoding/json/stream.go
@@ -204,7 +204,7 @@ func (enc *Encoder) Encode(v any) error {
 	}
 
 	e := newEncodeState()
-	defer encodeStatePool.Put(e)
+	defer cacheEncodeState(e)
 
 	err := e.marshal(v, encOpts{escapeHTML: enc.escapeHTML})
 	if err != nil {